Move `FCircuit` impls to `utils.rs` and add `DummyCircuit`
